Hi gents,
The jQuery specs on prop go into very specific detail on why prop() should be used for setting "checked" and "selected" values, and why checked="checked" is a static attribute that should not be used.

Excerpt:

Nevertheless, the most important concept to remember about the checked attribute is that it does not correspond to the checked property. The attribute actually corresponds to the defaultChecked property and should be used only to set the initial value of the checkbox. The checked attribute value does not change with the state of the checkbox, while the checked property does. Therefore, the cross-browser-compatible way to determine if a checkbox is checked is to use the property:
if ( elem.checked )
if ( $(elem).prop("checked") )
if ( $(elem).is(":checked") )

This is just a quick PR to transfer usages correctly onto prop. All test cases passed. Downloaded the lastest repo as of today.