Not accepting the element on the selector when the attribute is blank
The default method to find elements on the code was:
$('[name]', rootEl);
But this can end up selecting elements where the attribute exists but is blank, and this makes no sense on data binding.
With this fix this will not happen anymore.
The main drive to fix this is to run tests on Node, using JSDom and NPM's jquery. There is a bug on Node's jquery where the attribute selector does not work as expected and always selects all elements, regardless if they have the attribute or not. This fixes that as well.
Because of this problem, on Node nothing would work unless the elements  being bound where direct children of the RootEl.
This fix will allow the ModelBinder to work on the browser and on Node.